parser/requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
cachetools==6.1.0
catalogue==2.0.10
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cryptography==45.0.5
cymem==2.0.11
dateparser==1.2.2
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
Flask==3.1.1
flask-cors==6.0.1
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.0
langdetect==1.0.9
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
narwhals==1.46.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pdfminer.six==20250506
pillow==11.3.0
preshed==3.0.10
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.46.1
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
